+++ PersonalFinanceTracker/Models/Savings.cs
@@ -5,21 +5,14 @@ namespace PersonalFinanceTracker.Models
     public class Savings
     {
         public string Id { get; set; } = Guid.NewGuid().ToString("N");
-
         public string UserId { get; set; }
-
         public string Title { get; set; }
-
         public DateTime Date { get; set; }
-
         [Display(Name = "Target Amount")]
-        public float TargetAmount { get; set; }
-
+        public decimal TargetAmount { get; set; }
         [Display(Name = "Current Amount")]
-        public float CurrentAmount { get; set; }
-
+        public decimal CurrentAmount { get; set; }
         public DateTime Deadline { get; set; }
-
         public string Status { get; set; }
     }
 }
